renamed component to paymentMethodForm
