+++ BeaconAndLoaves/ClientApp/src/App/App.js
@@ -12,9 +12,7 @@ import Register from '../components/pages/Register/Register';
 import Home from '../components/pages/Home/Home';
 import Profile from '../components/pages/Profile/Profile';
 import Properties from '../components/pages/Properties/Properties';
-import PaymentMethod from '../components/pages/PaymentMethod/PaymentMethod';
 import SinglePaymentMethodScreen from '../components/pages/SinglePaymentMethod/SinglePaymentMethod';
-import PaymentMethodBTC from '../components/pages/PaymentMethod/PaymentMethodBTC';
 import LightHouses from '../components/pages/LightHouses/LightHouses';
 import SiloNuclears from '../components/pages/SiloNuclears/SiloNuclears';
 import LightHouseDetail from '../components/pages/LightHouseDetail/LightHouseDetail';
@@ -100,9 +98,7 @@ export default class App extends Component {
                   <PrivateRoute path='/register' exact component={Register} authed={this.state.authed} />
                   <PrivateRoute path="/home" component={Home} authed={this.state.authed}/>
                   <PrivateRoute exact path="/profile" component={Profile} authed={this.state.authed}/>
-                  <PrivateRoute path="/paymentMethod" component={PaymentMethod} authed={this.state.authed}/>
                   <PrivateRoute path="/SinglePaymentMethodScreen" component={SinglePaymentMethodScreen} authed={this.state.authed}/>
-                  <PrivateRoute path="/paymentMethodBTC" component={PaymentMethodBTC} authed={this.state.authed}/>
                   <PrivateRoute exact path="/properties" component={Properties} authed={this.state.authed}/>
                   <PrivateRoute path="/properties/lightHouses" component={LightHouses} authed={this.state.authed}/>
                   <PrivateRoute exact path="/lightHouses/:id" authed={this.state.authed} component={LightHouseDetail}/>
