+++ BeaconAndLoaves/ClientApp/src/components/pages/ConfirmationModal/ConfrimationModal.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  Button,
   Modal,
   ModalHeader,
   ModalBody,
