+++ BeaconAndLoaves/ClientApp/src/components/pages/PaymentMethodForm/PaymentMethodBTC.js
